\name{confintRob}
\alias{confintRob}
\title{Compute robust confidence interval}
\usage{
  confintRob(object, parm, level = 0.95, vcov = NULL, ...)
}
\arguments{
  \item{object}{}

  \item{parm}{}

  \item{level}{}

  \item{vcov}{}

  \item{...}{}
}
\value{
  confint
}
\description{
  Compute robust confidence interval
}
\author{
  ahmadou
}

